% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_zip_ref}
\alias{city_zip_ref}
\title{OC Reference}
\format{
An object of data.frame.
A data frame with 91 rows and 2 columns:
\describe{
  \item{Zip}{Five digit zip code}
  \item{City}{City name}
}
}
\usage{
city_zip_ref
}
\description{
Data.frame of OC zip codes and city names.
}
\keyword{datasets}
