% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_date.R
\name{excel_date}
\alias{excel_date}
\title{Convert Numeric Dates in Excel}
\usage{
excel_date(x)
}
\arguments{
\item{x}{Numeric date in Excel, class numeric.}
}
\value{
Date.
}
\description{
Convert Numeric Dates in Excel
}
\examples{
excel_date(42564)
}
