% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{expand_x}
\alias{expand_x}
\title{Add Length to X Scale}
\usage{
expand_x(x)
}
\arguments{
\item{x}{Numeric, how much to add to x-scale. Depends on what class x-scale is e.g. date, continuous.}
}
\value{
X-scale expanded to the right by x.
}
\description{
Add Length to X Scale
}
