% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{extract_zip}
\alias{extract_zip}
\title{Extract Zip Code}
\usage{
extract_zip(address, starts_with = 9)
}
\arguments{
\item{address}{Character}

\item{starts_with}{Numeric, starting number of local zip codes.}
}
\value{
Character, length of five.
}
\description{
Extract five digit zip code from single line address.
}
\examples{
extract_zip("1234 Main Street, Santa Ana, CA 92706")
}
