% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Convert NA Numeric Cells to Zero}
\usage{
na_to_zero(df)
}
\arguments{
\item{df}{Input data frame, usually following dplyr::count.}
}
\value{
Data frame with NAs replaced with zero.
}
\description{
Convert NA Numeric Cells to Zero
}
\examples{
df <- data.frame(loc = c("A","B","C"), n = c(5,NA,6))
na_to_zero(df)
}
