% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color.R
\name{scale_color_apollo}
\alias{scale_color_apollo}
\title{Apollo Color Scale}
\usage{
scale_color_apollo(name, n, reverse = FALSE, ...)
}
\arguments{
\item{name}{Current options - Bright, Muted, Oktabe Ito.}

\item{n}{Number of colors needed.}

\item{reverse}{Logical, default set to FALSE.}

\item{...}{Other arguments for ggplot2::discrete_scale.}
}
\value{
Color scale.
}
\description{
Apollo Color Scale
}
