% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{scale_percent}
\alias{scale_percent}
\title{Add Percent to Y-Axis Labels}
\usage{
scale_percent(x, ..., scale = TRUE)
}
\arguments{
\item{x}{Input list, leave blank.}

\item{...}{Other arguments from base::format.}

\item{scale}{Logical, default to TRUE. If FALSE, multiples labels by 100 to get percentage.}
}
\value{
Y-axis in ggplot2 formatted with percentage symbol.
}
\description{
Add Percent to Y-Axis Labels
}
