% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tbl}
\alias{tbl}
\title{Frequency Table}
\usage{
tbl(df, ..., scaled = TRUE, digits = 1)
}
\arguments{
\item{df}{Input data.frame}

\item{...}{Variables you want to summarize. If one variable, returns n and percent. If 2 or more, uses tidyr::pivot_wider and only returns n.}

\item{scaled}{Is percent multiplied by 100 or returned as fraction.}

\item{digits}{Rounding for percent.}
}
\value{
Frequency tbl.
}
\description{
Frequency Table
}
