% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tbl_totals}
\alias{tbl_totals}
\title{Column or Row Totals}
\usage{
tbl_totals(df, loc = c("col", "row", "both"), name = "Total")
}
\arguments{
\item{df}{Input data, usually follows dplyr::count.}

\item{loc}{Do you want to sum the row, column, or both.}

\item{name}{Name of summed celled. Defaults to total.}
}
\value{
Table with added totals.
}
\description{
Column or Row Totals
}
\examples{
df <- data.frame(loc = c("A","B","C"), n = c(5,10,6))
tbl_totals(df)
}
