% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{contains_columns}
\alias{contains_columns}
\title{Find Columns that Contain String}
\usage{
contains_columns(df, contains)
}
\arguments{
\item{df}{Data.frame}

\item{contains}{Search terms, search multiple patterns using |.}
}
\value{
Vector, charachter.
}
\description{
Find Columns that Contain String
}
