% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tbl_percentage}
\alias{tbl_percentage}
\title{Column Percentages}
\usage{
tbl_percentage(df, loc = c("row", "col"), digits = 1)
}
\arguments{
\item{df}{Input data, usually follows dplyr::count.}

\item{loc}{Either row or col. Where do you want the percentage to go.}

\item{digits}{For rounding.}
}
\value{
Table with cells as percentages of column.
}
\description{
Column Percentages
}
\examples{
df <- data.frame(loc = c("A","B","C"), n = c(5,10,6))
tbl_percentage(df)
}
